Allow deeper zoom out in compare screen
